pyproject.toml
[build-system]
requires = ["setuptools>=72.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "factor"
version = "0.1.0"
description = "Agentic AI legal due diligence platform — autonomous agents built on AWS Strands Agents SDK and Bedrock AgentCore"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "A Taylor" },
]
keywords = [
    "agentic-ai",
    "strands-agents",
    "bedrock-agentcore",
    "legal-tech",
    "ai-agents",
    "due-diligence",
    "contract-analysis",
    "rag",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "boto3>=1.35.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "datasets>=3.0.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
    "openpyxl>=3.1.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "moto>=5.0.0",
    "pytest-cov>=5.0.0",
]
infra = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/ATaylorAerospace/Factor-AI"
Repository = "https://github.com/ATaylorAerospace/Factor-AI"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
